Move more code around and remove hard-coded keys.
