+++ crates/core/Cargo.toml
@@ -141,6 +141,7 @@ indoc = "2.0.7"
 pglite-oxide = "0.5.1"
 quoted_printable = "0.5.2"
 rcgen = "0.14.10"
+rsa = "0.10.0-rc.15"
 schemars = "1.2.2"
 temp-dir = "0.2.0"
 tokio-metrics = "0.5.2"
