+++ crates/core/src/auth/create_external_user.rs
